# openapi/openapi.yaml
components:
  schemas:
...
    handlers.CreateOrganizationRequest:
      properties:
        admin_display_name:
          description: AdminDisplayName 是首个 org_admin 的显示名。
          type: string
        admin_password:
          description: AdminPassword 是首个 org_admin 的初始密码，service 层立即 hash 后写库。
          type: string
        admin_username:
          description: AdminUsername 是随组织创建的首个 org_admin 账号名。
          type: string
        code:
          description: Code 是组织登录标识，创建后不可修改。
          type: string
        contact_name:
          description: ContactName 是业务联系人姓名，可为空。
          type: string
        contact_phone:
          description: ContactPhone 是业务联系人电话，可为空；不参与权限或登录校验。
          type: string
        credit_warning_threshold:
          description: CreditWarningThreshold 是组织余额预警阈值；nil 表示不启用余额预警或保持预警关闭。
          type: integer
        enabled_models:
          description: EnabledModels 是该组织允许在 manager 内选择的模型列表；new-api token 不使用该字段限权。
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          description: Name 是组织展示名，也是平台管理员列表中识别租户的主字段。
          type: string
        remark:
          description: Remark 是平台管理员维护的内部备注。
          type: string
      required:
      - admin_display_name
      - admin_password
      - admin_username
      - code
      - name
      type: object
    handlers.ErrorResponse:
      properties:
        code:
          description: Code 是稳定的接口契约标识，一旦发布只增不改。
          example: APP_NOT_FOUND
          type: string
        message:
          description: Message 是面向前端展示的安全错误文案。
          example: 应用不存在
          type: string
      type: object
...
    handlers.OrganizationRequest:
      properties:
        contact_name:
          description: ContactName 是业务联系人姓名，可置空。
          type: string
        contact_phone:
          description: ContactPhone 是业务联系人电话，可置空。
          type: string
        credit_warning_threshold:
          description: CreditWarningThreshold 是组织余额预警阈值；nil 表示清空或未设置预警阈值。
          type: integer
        enabled_models:
          description: |-
            EnabledModels 是该组织允许在 manager 内选择的模型列表；new-api token 不使用该字段限权。
            nil 表示本次只更新基础资料并保留原模型列表；显式空数组会由 service 拒绝。
          items:
            type: string
          type: array
          uniqueItems: false
        name:
          description: Name 是组织展示名；更新时仍必填，避免空名称进入前端列表。
          type: string
        remark:
          description: Remark 是平台管理员维护的内部备注，可置空。
          type: string
      required:
      - name
      type: object
...
    service.OrganizationResult:
      properties:
        admin_username:
          description: AdminUsername 是组织首个可用管理员账号名，用于平台管理员复制登录信息。
          type: string
        code:
          description: Code 是组织登录标识，用于组织用户登录时定位租户。
          type: string
        contact_name:
          description: ContactName 是业务联系人姓名。
          type: string
        contact_phone:
          description: ContactPhone 是业务联系人电话。
          type: string
        credit_warning_threshold:
          description: CreditWarningThreshold 是组织余额预警阈值。
          type: integer
        enabled_models:
          description: EnabledModels 是组织在 manager 层允许创建实例时选择的模型列表。
          items:
            type: string
          type: array
          uniqueItems: false
        id:
          description: ID 是 manager 组织 UUID。
          type: string
        name:
          description: Name 是组织展示名。
          type: string
        newapi_user_id:
          description: NewAPIUserID 是组织在 new-api 侧的用户 ID，缺失时充值和用量接口不可用。
          type: string
        remark:
          description: Remark 是平台管理员维护的内部备注。
          type: string
        status:
          description: Status 是组织状态，active / disabled 决定成员是否可登录。
          type: string
      type: object
...
    service.QuotaSeries:
      properties:
        scope:
          description: Scope 标识配额序列所属维度，当前为 organization 或 platform。
          type: string
        scope_id:
          description: ScopeID 是组织维度的 manager org UUID；平台维度为空。
          type: string
        updated_at:
          description: UpdatedAt 是 manager 代理完成时刻，不代表 new-api 内部采集时间。
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 登录后获得的 JWT access token，前缀 "Bearer "。
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: 基于 Hermes Agent runtime 的多组织管理后台 API。
  title: Agent Runtime Manager API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /organizations:
    get:
      description: 平台管理员获取所有组织；org_admin 只能看到自己所属组织
      parameters:
      - description: 每页条数（默认 50）
        in: query
        name: limit
        schema:
          type: integer
      - description: 分页偏移（默认 0）
        in: query
        name: offset
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  items:
                    $ref: '#/components/schemas/service.OrganizationResult'
                  type: array
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 组织列表
      tags:
      - organizations
    post:
      description: 平台管理员创建新组织，并同步在 new-api 侧完成账户 provisioning
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.CreateOrganizationRequest'
                description: 创建组织请求
                summary: body
        description: 创建组织请求
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OrganizationResult'
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 创建组织
      tags:
      - organizations
  /organizations/{orgId}:
    get:
      description: 按 orgId 获取单个组织信息
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OrganizationResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 组织详情
      tags:
      - organizations
    patch:
      description: 更新组织名称、联系人、备注等基础字段
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/handlers.OrganizationRequest'
                description: 更新组织请求
                summary: body
        description: 更新组织请求
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OrganizationResult'
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 更新组织
      tags:
      - organizations
...
  /organizations/{orgId}/disable:
    post:
      description: 将组织状态设为 disabled，成员登录时将被拒绝
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OrganizationResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 禁用组织
      tags:
      - organizations
  /organizations/{orgId}/enable:
    post:
      description: 将组织状态从 disabled 恢复为 active
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.OrganizationResult'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: 启用组织
      tags:
      - organizations
...
  /usage/organizations/{orgId}:
    get:
      description: 返回指定组织在时间窗口内的按日 quota 消耗（QuotaSeries 格式）
      parameters:
      - description: 组织 ID
        in: path
        name: orgId
        required: true
        schema:
          type: string
      - description: 起始时间（Unix 秒）
        in: query
        name: since
        schema:
          type: integer
      - description: 结束时间（Unix 秒）
        in: query
        name: until
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  $ref: '#/components/schemas/service.QuotaSeries'
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/handlers.ErrorResponse'
          description: Service Unavailable
      security:
      - BearerAuth: []
      summary: 组织用量统计
      tags:
      - usage
...
servers:
- url: /api/v1